% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{assay2mtx}
\alias{assay2mtx}
\title{Extract expression matrix from Seurat object or matrix}
\usage{
assay2mtx(
  object,
  assay = NULL,
  slot = "counts",
  min.cells = 3,
  min.features = 1,
  update = F
)
}
\arguments{
\item{object}{A Seurat object or matrix containing expression data}

\item{assay}{A character string specifying the assay name to extract expression data from the Seurat object}

\item{slot}{A character string specifying the slot name to extract expression data from the Seurat object}

\item{min.cells}{An integer specifying the minimum number of cells required to keep genes in the expression matrix}

\item{min.features}{An integer specifying the minimum number of features required to keep genes in the expression matrix}
}
\value{
A matrix containing the expression data, with gene names as row names and cell barcodes as column names
}
\description{
This function extracts the expression matrix from either a Seurat object or a matrix object. If a Seurat object is provided, the function can optionally specify the assay and slot to be used for extracting the expression matrix. If a matrix is provided, the function will create a Seurat object with the provided matrix and default metadata fields. The function can also specify the minimum number of cells and features required to keep genes in the expression matrix.
}
\examples{
expmtx <- assay2mtx(object = seurat_object, assay = "RNA", slot = "counts", min.cells = 10, min.features = 5)
expmtx <-assay2mtx(object = expression_matrix, min.cells = 10, min.features = 5)

}
\author{
Qiong Zhang
}
