% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{buildPesudoBinMatrix}
\alias{buildPesudoBinMatrix}
\title{Construct the pesudo binary matrix}
\usage{
buildPesudoBinMatrix(clusterIDs, combine = FALSE, value = 3)
}
\arguments{
\item{clusterIDs}{a vector contains clusterID for cells}

\item{value}{fill in value}
}
\value{
pesudo binmatrix matrix
}
\description{
Construct the pesudo binary matrix
}
\author{
Qiong Zhang
}
