% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{calGeneFea}
\alias{calGeneFea}
\title{calculate the feature score of genes}
\usage{
calGeneFea(ginc, binmat, scale = F)
}
\arguments{
\item{ginc}{gene expression probability}

\item{binmat}{geneXcell binary matrix}

\item{scale}{scale the scores between groups}
}
\value{
feature score of genes
}
\description{
calculate the feature score of genes
}
\author{
Qiong Zhang
}
