% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{calLLS}
\alias{calLLS}
\alias{calLLRSim}
\title{Calculate log-likelihood similarity for two vectors}
\usage{
calLLRSim(a, b, n)
}
\arguments{
\item{a}{vector a}

\item{b}{vector b}
}
\value{
log-likelihood similarity between two vectors
}
\description{
Calculate log-likelihood similarity for two vectors
}
\author{
Qiong Zhang
}
