% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{calMutInfo}
\alias{calMutInfo}
\title{Calculate mutual information for two vectors}
\usage{
calMutInfo(a, b)
}
\arguments{
\item{a}{vector a}

\item{b}{vector b}
}
\value{
mutual information between two vectors
}
\description{
Calculate mutual information for two vectors
}
\author{
Qiong Zhang
}
