% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{calProb4C}
\alias{calProb4C}
\title{calculate prior probabilities for clusters}
\usage{
calProb4C(x)
}
\arguments{
\item{x}{vector}
}
\value{
probabilities of clusters
}
\description{
calculate prior probabilities for clusters
}
\author{
Qiong Zhang
}
