% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{calProb4G}
\alias{calProb4G}
\title{calculate prior probabilities for genes}
\usage{
calProb4G(mat)
}
\arguments{
\item{mat}{expression matrix}
}
\value{
probabilities for genes
}
\description{
calculate prior probabilities for genes
}
\author{
Qiong Zhang
}
