% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{calProbCIG}
\alias{calProbCIG}
\title{calculate probabilities of genes in given condition}
\usage{
calProbCIG(binmat, probG, ProbC)
}
\arguments{
\item{mat}{expression matrix}
}
\value{
probabilities of genes in given condition
}
\description{
calculate probabilities of genes in given condition
}
\author{
Qiong Zhang
}
