% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{calProbGIC}
\alias{calProbGIC}
\title{calculate probabilities genes in clusters}
\usage{
calProbGIC(mat, cluProb, cluID, cpu = 1)
}
\arguments{
\item{mat}{expression matrix}

\item{cluProb}{cluster}
}
\value{
probabilities of genes in clusters
}
\description{
calculate probabilities genes in clusters
}
\author{
Qiong Zhang
}
