% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{calTFIDF}
\alias{calTFIDF}
\title{Calculate TFIDF for clusters}
\usage{
calTFIDF(condMat)
}
\arguments{
\item{x}{a}

\item{b}{vector b}
}
\value{
calculate TFIDF for clusters
}
\description{
Calculate TFIDF for clusters
}
\author{
Qiong Zhang
}
