% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{clusidQC}
\alias{clusidQC}
\title{QC for clusterIDs}
\usage{
clusidQC(mat, clusterIDs)
}
\arguments{
\item{mat}{gene expression matrix}

\item{clusterIDs}{a vector contains clusterID for cells}
}
\value{
validated clusterIDs
}
\description{
QC for clusterIDs
}
\author{
Qiong Zhang
}
