% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{expMat2Bin}
\alias{expMat2Bin}
\title{Transform expression matrix to binary type based on give mode or customize value}
\usage{
expMat2Bin(mat, pattern = "min", cutoff = NULL)
}
\arguments{
\item{mat}{vector}

\item{pattern}{transform pattern}
}
\value{
binary matrix
}
\description{
Transform expression matrix to binary type based on give mode or customize value
}
\author{
Qiong Zhang
}
