% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{ginirev}
\alias{ginirev}
\title{Calculate gini specific score for markers}
\usage{
ginirev(x, rev = T)
}
\arguments{
\item{x}{vector of gene}
}
\value{
feature-specific gini scores.
}
\description{
Calculate gini specific score for markers
}
\author{
Qiong Zhang
}
