% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{markerREC}
\alias{markerREC}
\title{Recommend marker genes for expression matrix or Seurat object}
\usage{
markerREC(
  obj,
  assay = "RNA",
  slot = "data",
  binmethod = "min",
  expR = 0.1,
  cpu = 1,
  marker = 20,
  clusterIDs = NULL
)
}
\arguments{
\item{obj}{Seurat object or expression matrix}

\item{binmethod}{Method used to binarize the matrix}

\item{expR}{The proportion of feature expressed cells in a cluster.}

\item{cpu}{How many cores will be used for detection of markers}

\item{marker}{The number of markers returned.}

\item{clusterIDs}{The cluster ID for each cell or the column name which contains cluster ID information in Seurat object.}
}
\value{
data.frame of marker features in each cluster.
}
\description{
Recommend marker genes for expression matrix or Seurat object
}
\examples{
marker.mr <- markerREC(pbmc3k.final)

}
\author{
Qiong Zhang
}
