% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{probEnt}
\alias{probEnt}
\title{Calculate the entropy of probability vector}
\usage{
probEnt(x)
}
\arguments{
\item{x}{probability vector}
}
\value{
the entropy of probability vector
}
\description{
Calculate the entropy of probability vector
}
\author{
Qiong Zhang
}
