% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{scaleZO}
\alias{scaleZO}
\title{rescale a vector between 0 (lowest) and 1 (highest)}
\usage{
scaleZO(x)
}
\arguments{
\item{x}{vector}
}
\value{
rescale vector with value between 0 and 1
}
\description{
rescale a vector between 0 (lowest) and 1 (highest)
}
\author{
Qiong Zhang
}
