% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerREC.R
\name{shannoEnt}
\alias{shannoEnt}
\title{Calculate the shanno entropy for vector}
\usage{
shannoEnt(x)
}
\arguments{
\item{x}{vector}
}
\value{
the entropy value
}
\description{
Calculate the shanno entropy for vector
}
\author{
Qiong Zhang
}
